improve core layout
